Put Ares functions in a separate file
