+++ src/Misc/AresData.h
@@ -3,7 +3,6 @@
 #include <string>
 #include <unordered_map>
 #include <Windows.h>
-#include <ASMMacros.h>
 
 class TechnoClass;
 class TechnoTypeClass;
@@ -24,8 +23,6 @@ public:
 private:
 	// timestamp bytes for each version
 	static const std::unordered_map<DWORD, Version> AresTimestampBytes;
-	// storage for absolute addresses of functions (module base + offset)
-	static std::unordered_map<std::string, DWORD> AresFunctionOffsetsFinal;
 
 	static void GetGameModulesBaseAddresses();
 
@@ -36,37 +33,12 @@ public:
 
 	// offsets of function addresses for each version
 	static const std::unordered_map<std::string, std::vector<DWORD>> AresFunctionOffsets;
+	// storage for absolute addresses of functions (module base + offset)
+	static std::unordered_map<std::string, DWORD> AresFunctionOffsetsFinal;
 	// numeric id of currently used version, zero-indexed, -1 is unknown or missing
 	static Version AresVersion;
 	// is Ares detected and version known?
 	static bool CanUseAres;
 
 	static void Init();
-
-#define JMP_ARES(name) const auto x = AresFunctionOffsetsFinal[name]; JMP(x)
-#define JMP_THIS_ARES(name) const auto x = AresFunctionOffsetsFinal[name]; JMP_THIS(x)
-
-	// here be known Ares functions
-	static bool ConvertTypeTo(TechnoClass* pFoot, TechnoTypeClass* pConvertTo)
-	{
-		JMP_ARES("ConvertTypeTo");
-	}
-
-	static void SpawnSurvivors(FootClass* const pThis, TechnoClass* const pKiller, const bool Select, const bool IgnoreDefenses)
-	{
-		JMP_ARES("SpawnSurvivors");
-	}
-
-	static int HasFactory(int buffer, HouseClass* pOwner, TechnoTypeClass* pType, bool skipAircraft, bool requirePower, bool checkCanBuild, bool unknown)
-	{
-		JMP_ARES("HasFactory");
-	}
-
-	static bool CanBeBuiltAt(DWORD pTechnoTypeExt, BuildingTypeClass* pBuildingType)
-	{
-		JMP_THIS_ARES("CanBeBuiltAt");
-	}
-
-#undef JMP_ARES
-#undef JMP_THIS_ARES
 };
